+++ src/main/java/com/formssafe/domain/question/repository/DescriptiveQuestionRepository.java
@@ -3,5 +3,5 @@ package com.formssafe.domain.question.repository;
 import com.formssafe.domain.question.entity.DescriptiveQuestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DescriptiveQuestionRepository extends JpaRepository<DescriptiveQuestion, Integer> {
+public interface DescriptiveQuestionRepository extends JpaRepository<DescriptiveQuestion, Long> {
 }
