+++ src/main/java/com/formssafe/domain/reward/repository/RewardRecipientRepository.java
@@ -3,5 +3,5 @@ package com.formssafe.domain.reward.repository;
 import com.formssafe.domain.reward.entity.RewardRecipient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RewardRecipientRepository extends JpaRepository<RewardRecipient, Integer> {
+public interface RewardRecipientRepository extends JpaRepository<RewardRecipient, Long> {
 }
